requirements.txt
streamlit==1.32.0
torch==2.2.0
pillow==10.2.0
numpy==1.24.3
opencv-python==4.9.0.80
scikit-image==0.22.0